backend/pyproject.toml
[project]
name = "class-trader-backend"
version = "0.1.0"
description = "LLM-driven autonomous trading platform — backend"
requires-python = ">=3.12"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "websockets>=12.0",

    # Database
    "sqlalchemy[asyncio]>=2.0.30",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",

    # LLM
    "anthropic>=0.40.0",
    "instructor>=1.5.0",

    # Data validation
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",

    # Broker
    "alpaca-py>=0.30.0",

    # Market data
    "finnhub-python>=2.4.0",
    "requests>=2.32.0",
    "aiohttp>=3.10.0",

    # Reddit / retail sentiment
    "asyncpraw>=7.7.0",
    "praw>=7.7.0",

    # Scheduling
    "apscheduler>=3.10.0",

    # Notifications
    "apprise>=1.9.0",

    # Utilities
    "python-dotenv>=1.0.0",
    "jinja2>=3.1.0",
    "httpx>=0.27.0",
    "pandas>=2.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.mypy]
python_version = "3.12"
strict = true
